AntennaPattern

Documentation for AntennaPattern.

AntennaPattern.DEFAULTS_1D

AntennaPattern.antennaPattern3D

AntennaPattern.cart2sph

AntennaPattern.cartesianPattern

AntennaPattern.createSurface

AntennaPattern.polarPattern

AntennaPattern.preprocess_CST_3D_ASCII_file

AntennaPattern.setBackend1D

AntennaPattern.setBackend3D

AntennaPattern.sph2cart

AntennaPattern.@STANDARD_CST_3D_PATTERN_COLUMNS

AntennaPattern.DEFAULTS_1D — Constant

Defaut values for the 1D antenna pattern plot

AntennaPattern.antennaPattern3D — Method

antennaPattern3D

Plots the 3D antenna pattern.

Args:

X::Matrix{Float64}: x coordinate, [N x M]

Y::Matrix{Float64}: y coordinate, [N x M]

Z::Matrix{Float64}: z coordinate, [N x M]

RN::Matrix{Float64}: normalized data, [N x M]

kwargs...: keyword arguments

width::Int: width of the plot, default: 800

height::Int: height of the plot, default: 800

xlabel::String: x label, default: "X"

ylabel::String: y label, default: "Y"

zlabel::String: z label, default: "Z"

title::String: title of the plot, default: "3D Antenna Pattern"

show_grid::Bool: show grid, default: false

show_axis::Bool: show axis, default: false

legend::Bool: show legend, default: true

color_scale_plotly::String: color scale for PlotlyJS, default: "Rainbow"

color_scale_pyplot::Symbol: color scale for PyPlot, default: :jet

aaxis_show::Bool: show axis, default: true

aaxis_theta::Float64: theta angle of the axis, default: 90.0

aaxis_phi::Float64: phi angle of the axis, default: 0.0

aaxis_color::Symbol: color of the axis, default: :blue

aaxis_linewidth::Float64: line width of the axis, default: 3.0

aaxis_linestyle::Symbol: line style of the axis, default: :auto

aaxis_label::String: label of the axis, default: ""

aaxis_axis_multiplier::Float64: axis multiplier, default: 1.3

Returns:

p: plot

AntennaPattern.cart2sph — Method

cart2sph

Transforms carthesian coordinates to spherical coordinates

INPUTS:

x: x coordinate, Real

y: y coordinate, Real

z: z coordinate, Real

OUTPUTS

r: r coordinate in spherical coordinates, Real

θ: theta coordinate in spherical coordinates, Real

φ: phi coordinate in spherical coordinates, Real

type: [x, y, z]

AntennaPattern.cartesianPattern — Function

cartesianPattern

Converts spherical pattern to cartesian pattern and sets its dynamic range.

Args:

θ::Union{Matrix{Float64}, Matrix{Real}}: θ coordinate in radians, [N x M]

φ::Union{Matrix{Float64}, Matrix{Real}}: φ coordinate in radians, [N x M]

r::Union{Matrix{Float64}, Matrix{Real}}: r, [N x M]

dymanicMinimum::Real: dymanicMinimum, default: 0, sets the dynamic range of the data to [dymanicMinimum, max(data)], must be lesser or equal to 0

Returns:

x: x coordinate, [N x M]

y: y coordinate, [N x M]

z: z coordinate, [N x M]

r_norm: normalized r, [N x M]

AntennaPattern.createSurface — Method

createSurface

Creates a surface from the given data.

Args:

θ::Union{Vector{Real}, Vector{Float64}}: θ coordinate in radians

φ::Union{Vector{Real}, Vector{Float64}}: φ coordinate in radians

data::Union{Vector{Float64}, Vector{Real}, Matrix{Real}, Matrix{Float64}}: data

repeatFirstφ::Bool: repeat first φ value at the end, default: false

θf::Symbol: θ format, default: :rad

φf::Symbol: φ format, default: :rad

Possible combinations of θ, φ and data:

θ::Vector, φ::Vector, data::Vector - data must be of size (θn*φn)

θ::Vector, φ::Vector, data::Matrix - data must be of size (θn, φn)

Returns:

θr: θ coordinate in radians, [N x M]

φr: φ coordinate in radians, [N x M]

r: r, [N x M]

AntennaPattern.polarPattern — Method

polarPattern

Polar plot of an antenna pattern.

Arguments

Constant::Symbol: The constant angle to plot the antenna pattern at. Must be either :Theta or :Phi.

CutAngle::Real: The angle at which to plot the antenna pattern. Must be between 0 and 360.

θ::Union{Vector{Real}, Vector{Float64}}: The θ values of the antenna pattern (degrees).

φ::Union{Vector{Real}, Vector{Float64}}: The φ values of the antenna pattern (degrees).

data::Union{Vector{Real}, Vector{Float64}}: The data values of the antenna pattern.

repeatFirst::Bool: Whether to repeat the first value of the antenna pattern at the end. Defaults to true.

kwargs...: Keyword arguments to pass to the plot function.

kwargs

width::Int: The width of the plot. Defaults to 600.

height::Int: The height of the plot. Defaults to 500.

label::String: The label of the plot. Defaults to "Antenna Pattern".

legend::Bool: Whether to show the legend. Defaults to true.

line_width::Float64: The width of the line. Defaults to 3.0.

color::Symbol: The color of the line. Defaults to :red.

title::String: The title of the plot. Defaults to "Antenna Pattern".

mainlobe::Bool: Whether to show the main lobe. Defaults to true.

mainlobe_color::Symbol: The color of the main lobe. Defaults to :blue.

mainlobe_linewidth::Float64: The width of the main lobe line. Defaults to 1.5.

mainlobe_label::String: The label of the main lobe. Defaults to "Main lobe".

rotation::Float64: The rotation of the plot in degrees. Defaults to 0.0.

Returns

p::Plot: The plot object.

AntennaPattern.preprocess_CST_3D_ASCII_file — Method

Preprocessing of CST 3D ASCII file

Generates a CSV file from a CST 3D ASCII file. The CST 3D ASCII file is a text file with the following format:

- ["Theta[deg.]", "Phi[deg.]", "|Dir.|[dBi]", "|Theta|[dBi]", "∠Theta[deg.]", "|Phi|[dBi]", "∠Phi[deg.]", "AR[dB]"]

Inputs:

- input_filename: path to the input file
- output_filename: path to the output file

Outputs:

- output_filename: processed file

AntennaPattern.setBackend1D — Method

setBackend1D

Sets the plotting engine. Available engines are:

:gr

AntennaPattern.setBackend3D — Method

setBackend3D

Sets the plotting engine. Available engines are:

:plotlyjs

:pyplot

AntennaPattern.sph2cart — Method

sph2Cart

Transforms spherical coordinates to carthesian coordinates

INPUTS

r: r coordinate in spherical coordinates, Real

θ: theta coordinate in spherical coordinates, Real

φ: phi coordinate in spherical coordinates, Real

OUTPUTS

x: x coordinate, Real

y: y coordinate, Real

z: z coordinate, Real

type: [x, y, z]

AntennaPattern.@STANDARD_CST_3D_PATTERN_COLUMNS — Macro

Standard CST 3D ASCII file header